% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{as_matrix}
\alias{as_matrix}
\title{Title
this function turns sparse matrix into norm matrix, especially for those matrixs which can't be changed by as.matrix()}
\usage{
as_matrix(mat)
}
\arguments{
\item{mat}{}
}
\value{

}
\description{
Title
this function turns sparse matrix into norm matrix, especially for those matrixs which can't be changed by as.matrix()
}
\examples{
i <- c(1,3:8);
j <- c(2,9,6:10);
x <- 7 * (1:7)
(A <- sparseMatrix(i, j, x = x))
A.norm <- as_matrix(A)

}
