% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sample_seurat}
\alias{sample_seurat}
\title{Sample Seurat Object
 @description  can provide a subset which are randomly select from a given Seurat Object with a very rate.}
\usage{
sample_seurat(sobj, sample_rate = 1, rep = F)
}
\arguments{
\item{sobj}{seurat object}

\item{sample_rate}{sample rate. default is 1}

\item{rep}{whether want repeated, default is False}
}
\value{

}
\description{
Sample Seurat Object
 @description  can provide a subset which are randomly select from a given Seurat Object with a very rate.
}
\examples{
data("pbmc3k")
sample_seurat(pbmc3k,sample_rate=0.5)

}
