% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{scrna_pre_process}
\alias{scrna_pre_process}
\title{Title}
\usage{
scrna_pre_process(sobj, reduce_method = "pca", batch_label = NULL, label = F)
}
\arguments{
\item{reduce_method}{choose from "pca","umap","tsne"}

\item{batch_label}{if there may be batch effects, use meta data label}

\item{label}{if output with label}

\item{s.obj}{seurat object}
}
\value{

}
\description{
Title
}
\examples{
data("pbmc3k")
scrna_pre_process(pbmc3k)

}
